Retirement UI - WIP New step to choose a toucan project to retire to
