+++ retirement-ui/src/components/searchableDropdown.tsx
@@ -1,34 +1,60 @@
-import {FC, useState} from "react";
+import {RefObject, useRef, useState, FocusEvent} from "react";
+import {FaTimesCircle} from "react-icons/fa";
 
-type Option = { name: string; id: string };
-export const SearchableDropdown: FC<{options: Option[], loading: boolean}> = ({ options, loading }) => {
+type Option<T> = { name: string; id: string } & T;
+type SearchableDropdownProps<T> = { options: Option<T>[], loading: boolean, select: (option: Option<T>) => void };
+
+export function SearchableDropdown<T>({ options, loading, select }: SearchableDropdownProps<T>) {
     const [search, setSearch] = useState('');
     const [isOpen, setIsOpen] = useState(false);
+    const containerRef: RefObject<HTMLDivElement> = useRef(null);
+
+    const handleBlur = (e: FocusEvent<HTMLDivElement, Element>) => {
+        // Check if the blur event is triggered by clicking on the dropdown itself.
+        // If not, then close the dropdown.
+        if (!containerRef.current?.contains(e.relatedTarget)) {
+            setIsOpen(false);
+        }
+    };
 
     const filteredOptions = options.filter(option =>
         option.name.toLowerCase().includes(search.toLowerCase())
     );
 
+    const clearInput = () => {
+        setSearch('');
+    };
+
     return (
-        <div className="relative">
-            <input
-                type="text"
-                className="border rounded p-2"
-                value={search}
-                onChange={e => setSearch(e.target.value)}
-                onFocus={() => setIsOpen(true)}
-                placeholder="Search..."
-            />
+        <div
+            ref={containerRef}  onBlur={handleBlur}>
+            <div className="relative inline-block">
+                <input
+                    type="text"
+                    className="border rounded p-2 pr-8 w-96"
+                    value={search}
+                    onChange={e => setSearch(e.target.value)}
+                    onFocus={() => setIsOpen(true)}
+                    placeholder="Search..."
+                />
+                {search && (
+                    <FaTimesCircle
+                        className="absolute right-2 top-3 text-gray-400 cursor-pointer"
+                        onClick={clearInput}
+                    />
+                )}
+            </div>
             {loading && <div className="text-gray-500 mb-2">Loading...</div>}
             {!loading && isOpen && (
-                <ul className="absolute z-10 mt-2 w-full border rounded shadow-md">
+                <ul className="absolute z-10 mt-2 w-full h-64 bg-white border rounded shadow-md overflow-auto">
                     {filteredOptions.map(option => (
                         <li
-                            key={option.id}
+                            key={option.name}
                             className="p-2 hover:bg-gray-200 cursor-pointer"
-                            onClick={() => {
+                            onMouseDown={() => {
                                 setSearch(option.name);
                                 setIsOpen(false);
+                                select(option);
                             }}
                         >
                             {option.name}
