+++ retirement-ui/src/hooks/holdingContract/useHoldingContractBalance.ts
@@ -1,13 +1,13 @@
-import {useState} from "react";
-import {PublicKey} from "@solana/web3.js";
 import {HOLDING_CONTRACT_ADDRESS, USDC_TOKEN_POLYGON} from "@/lib/constants";
 import {useBalance} from "wagmi";
+import {Address} from "abitype/src/abi";
 
-export const useHoldingContractBalance = (): { balance: bigint | undefined, loading: boolean, error: Error | null }  => {
+export const useHoldingContractBalance = (holdingContractAddress: Address | undefined): { balance: bigint | undefined, loading: boolean, error: Error | null }  => {
     const { data, isError, isLoading, error } = useBalance({
         address: HOLDING_CONTRACT_ADDRESS,
         token: USDC_TOKEN_POLYGON,
-        watch: true
+        watch: true,
+        enabled: !!holdingContractAddress
     })
 
     return {
